// src/payload-types.ts
export interface LegalPage {
  id: number;
  title: string;
  slug:
    | 'terms-and-conditions'
    | 'cookies-policy'
    | 'privacy-policy'
    | 'refund-policy'
    | 'editorial-policy'
    | 'sustainability-policy';
  summary?: string | null;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  updatedAt: string;
  createdAt: string;
}